Rename library/bundleSizeDiff -> library/bundleSize
The directory name predated the `flub generate bundleSizeDiff` -> `flub
check bundleSize` rename. Match the consumer-facing surface and drop the
"Diff" framing — the library does the diff, but also baseline fetching,
ADO interaction, artifact extraction, and path classification.
