+++ build-tools/packages/build-cli/src/library/bundleSize/utilities/getAllFilesInDirectory.ts
